requirements.txt
streamlit
pandas
plotly_express==0.4